+++ fltk-sys/src/menu.rs
@@ -370,7 +370,10 @@ extern "C" {
     pub fn Fl_Menu_Bar_set_text_color(arg1: *mut Fl_Menu_Bar, c: ::std::os::raw::c_uint);
 }
 extern "C" {
-    pub fn Fl_Menu_Bar_add_choice(arg1: *mut Fl_Menu_Bar, arg2: *const ::std::os::raw::c_char);
+    pub fn Fl_Menu_Bar_add_choice(
+        arg1: *mut Fl_Menu_Bar,
+        arg2: *const ::std::os::raw::c_char,
+    ) -> ::std::os::raw::c_int;
 }
 extern "C" {
     pub fn Fl_Menu_Bar_get_choice(arg1: *mut Fl_Menu_Bar) -> *const ::std::os::raw::c_char;
@@ -816,7 +819,7 @@ extern "C" {
     pub fn Fl_Menu_Button_add_choice(
         arg1: *mut Fl_Menu_Button,
         arg2: *const ::std::os::raw::c_char,
-    );
+    ) -> ::std::os::raw::c_int;
 }
 extern "C" {
     pub fn Fl_Menu_Button_get_choice(arg1: *mut Fl_Menu_Button) -> *const ::std::os::raw::c_char;
@@ -1238,7 +1241,10 @@ extern "C" {
     pub fn Fl_Choice_set_text_color(arg1: *mut Fl_Choice, c: ::std::os::raw::c_uint);
 }
 extern "C" {
-    pub fn Fl_Choice_add_choice(arg1: *mut Fl_Choice, arg2: *const ::std::os::raw::c_char);
+    pub fn Fl_Choice_add_choice(
+        arg1: *mut Fl_Choice,
+        arg2: *const ::std::os::raw::c_char,
+    ) -> ::std::os::raw::c_int;
 }
 extern "C" {
     pub fn Fl_Choice_get_choice(arg1: *mut Fl_Choice) -> *const ::std::os::raw::c_char;
@@ -1687,7 +1693,7 @@ extern "C" {
     pub fn Fl_Sys_Menu_Bar_add_choice(
         arg1: *mut Fl_Sys_Menu_Bar,
         arg2: *const ::std::os::raw::c_char,
-    );
+    ) -> ::std::os::raw::c_int;
 }
 extern "C" {
     pub fn Fl_Sys_Menu_Bar_get_choice(arg1: *mut Fl_Sys_Menu_Bar) -> *const ::std::os::raw::c_char;
@@ -1920,6 +1926,12 @@ extern "C" {
         data: *mut ::std::os::raw::c_void,
         flag: ::std::os::raw::c_int,
     ) -> ::std::os::raw::c_int;
+}
+extern "C" {
+    pub fn Fl_Menu_Item_set_shortcut(self_: *mut Fl_Menu_Item, shortcut: ::std::os::raw::c_int);
+}
+extern "C" {
+    pub fn Fl_Menu_Item_set_flag(self_: *mut Fl_Menu_Item, flag: ::std::os::raw::c_int);
 }
 extern "C" {
     pub fn Fl_mac_set_about(